[LFSIndexCache] Make leaf size extraction more generally useful
The leaf size extraction support used by LFSIndexCache can be useful in
different places, but is currently closely tied to the needs of the
LFSIndexCache.

This patch improves on the current situation by switching to an
iterator-based interface so we can handle different types of containers
and adds a wrapper function extract_lfs_leaf_sizes() to hide the
creation and application of the visitor.