[Constraints] Move NoConstraints to constraints/noconstraints.hh
Having the empty constraints implementation in
gridfunctionspace/utility.hh isn't very intuitive, so this patch moves
it to a new header noconstraints.hh in the constraints/ folder. I've
kept the old file with a deprecation message, but most users probably
won't notice because the file was automatically included for them by
gridfunctionspace.hh.