[Adaptivity] Deprecate versions of adapt_grid() that set an implicit mass matrix integration order
By default, adapt_grid() uses a second order integration rule for
calculating the element mass matrices. That rule isn't sufficient for
elements of order > 1, causing the subsequent matrix inversion to fail.

The resulting exception is probably really confusing to a user, so this
patch makes sure that users are aware of the problem and encourages them
to explicitly spefify the correct integration order.