[StationaryLinearProblemSolver] Code cleanup and performance improvement
This patch adds an option to keep the allocated matrix across
invocations of the solver to avoid rebuilding the pattern for the common
case where it doesn't change between calls to the solver.
The matrix is automatically created if it does not exist upon entering
the apply() method. By default, the matrix is now kept across calls to
apply(); it can be discarded manually by calling discardMatrix() and the
old behavior can be restored with setKeepMatrix(false).

The patch also cleans up the code (whitespace, formatting, variable
naming conventions).