[StationaryLinearProblemSolver] Add constructors that take a ParameterTree
Having an ever growing list of defaulted constructor arguments for
components like the StationaryLinearProblemSolver doesn't really scale
and can be really inconvenient if you only want to change argument
number 6 and have to provide values for all earlier arguments just to do
so.

Using a ParameterTree avoids that problem and also provides for an easy
way to specify those parameters from the outside at run time. The new
constructors have a detailed Doxygen comment with a table of all the
keys currently recognized by them.