[Bugfix][StationaryLinearProblemSolver] Fix calculation of reduction for linear backend
StationaryLinearProblemSolver calculated the desired reduction as passed
to the linear solver backend as the minimum of the reduction requested
by the user and of start_defect / min_defect.

This is wrong in two respects:

- It should use the maximum, as we only require one of the two
  criteria (relative reduction and absolute defect) to be met.

- The conversion from absolute defect to reduction is wrong, it has to
  be min_defect / start_defect. This problem was masked by the first
  one; otherwise the linear solver would almost always have converged in
  a single iteration.

The patch also adds a little explanation of the meaning of the
parameters reduction and min_defect to the doxygen comment of the new constructors.