Replace template parameter qorder by run-time parameter quadOrder
The quadrature order used by the local assembler used to be a template
parameter of the Poisson class.  This patch removes this template parameter
und replaces it by a run-time parameter of the constructor.  IMO, this
is easier to use, more flexible, and it is doubtful whether a compile-time
quadrature order would lead to run-time savings at all.