[function] introduce base class for all global discrete functions
all functionality of DiscreteGridViewFunction has been moved to
DiscreteGridViewFunctionBase. The actual implementation will be
determined by the LocalFunction. Traits collect all typedefs necessary
to export the correct interface. The DiscreteGridViewFunctionBase is
templated by the LocalFunction implementation. On the one-hand-side it
contradict the hierarchy introduced in dune-functions, on the otherhand
it avoids a lot of copy-and-past code.