[function] introduce base class for all local discerte functions
an in the case of global discrete functions all functionality of DiscreteGridViewFunction 
has been moved to DiscreteLocalGridViewFunctionBase. This base class should hold the
common members, storage, etc. It implements bind and unbind and hides most of the ugly
parts when dealing with PDELab local function spaces. The acutal implementation only 
implements the derivative method and the evaluation.