Add local operator that tests a given function against an FE space on the entire grid (aka lambda_volume)
Such an operator is helpful because it appears as the source term in the
weak form of many equations.  Also, it allows to compute the integrals
over all basis functions of an FE basis (by testing the constant 'one'
function), which is needed for energy minimization problems involving
superposition operators.