[RannacherTurekFEM] Bring RannacherTurekFEM up to date and provide dimension-independent class
The Rannacher-Turek element in dune-localfunctions was changed to work
in a dimension-independent way during the 2.3 release cycle. As the old
version of the element has been deprecated, we need to switch to the new
one.

We also grab the chance and make the FEM frontend class
dimension-independent as well, deprecating the old 2D version.