[Constraints] Track presence of non-Dirichlet constraints in constraints container
This patch extends the ConstraintsTransformation interface with a new
method containsNonDirichletConstraints(), making it possible for the
assembly process to exploit the common case of only having to
incorporate Dirichlet constraints.

The value of this flag is cached in the default implementation, so it is
important to only add new constraints by means of the method
import_local_transformation(), which takes care of updating the cached value.