[LFSIndexCache] Make the caching of constraints controllable via a run-time parameter
The LFSIndexCache currently insists on always caching the constraints
for the DOFs of the currently bound LFS; that behavior can only be
deactivated by passing an EmptyTransformation for the type of the
constraints container, as the cache is specialized for that case and
then completely disables all constraints handling.

However, this way of disabling constraints caching modifies the type of
the LFSIndexCache, which creates a problem for the GridOperator engines,
which need the precise type to instantiate local views for the LA
containers and which creates problems in the case of the combined
residual and jacobian engine of the OneStepGridOperator, where the
residual doesn't require constraints caching, but the jacobian does,
causing a compilation failure.

Moreover, it would be beneficial to control the caching at run-time, as
the full-blown constraints handling that requires caching is only needed
in the presence of non-Dirichlet constraints, something that can only be
queried at runtime.

So this patch adds a new runtime flag to the LFSIndexCache that can be
used to enable or disable constraints caching. It can be passed as a
third parameter to the constructor, but is optional right now and
defaults to true. It is also possible to query the state of the flag
using the new method constraintsCachingEnabled().