[GridOperator] Make constraints caching controllable at run time
This patch switches the query mechanism of the global assembler for the
constraints caching requirements from a compile time interface (static
const bool needs_constraints_caching) to a run time mehod
needsConstraintsCaching(cu,cv). This new method is passed the
constraints containers for both spaces to allow decisions based on the
actual structure of the constraints (e.g. presence of non-Dirichlet
constraints).

Moreover, the patch also changes the non-constrained LFSCache typedefs
in the local assembler to be identical to the constrained ones, as the
caching-decision now happens at run time. This change makes it possible
to use the residual engine in settings without constraints
caching (e.g. when directly calling residual() ) as well as in the
context of the combined residual / jacobian assembly of the
ExplicitGridOperator, where the jacobian might enforce constraints
caching, even though the residual does not need it.

Finally, the patch changes the default constraints caching policy of
several engines:

- The residual engine never required constraints caching, because the
  residual is constrained in a batch operation after assembly is
  done. For some reason, constraints caching was activated
  anyway (probably because of the issue with the combined residual /
  jacobian method of the OneStepGridOperator). Caching has been disabled
  for this engine.

- The jacobian engine now only enables constraints caching if there are
  non-Dirichlet constraints, as Dirichlet constraints are applied after
  assembly is done, just like the residual.

- The pattern engine also only enables caching in the presence of
  non-Dirichlet constraints.

WARNING: This patch modifies the internal API between global assembler
         and assembler engines! Any users with custom assembler engines
         need to add the method needsConstraintsCaching() to their
         implementations!