[GridOperator] Add new method scatter_jacobian() as front end for etadd() et. al.
The new method scatter_jacobian() on the local assembler should be
called instead of etadd(). It inspects the index caches to decide
if the expensive etadd() method has to be invoked. Otherwise, the local
jacobian will be written back without taking constraints into
consideration.

Moreover, the new method has a parameter symmetric_mode that enables
symmetric assembly of Dirichlet constraints.