[LocalOperator] Add new tag class DiagonalJacobian
Decorator base class for local operators with diagonal jacobian to avoid
the overhead of a dense local jacobian matrix in that case.
For high-order DG in 3D (5+), the local jacobian becomes really
large (e.g. 2MB for Q7), which completely thrashes the cache, destroying
scalability in the polynomial order. A number of follow-up patches will
use the information provided by this decorator to improve scaling in
that case.