[bugfix,pkg-config] Add library to pkg-config file.
Obviously this module ships a library but lacks it in the linker flags
for pkg-config. Therefore the build of modules that depend on an
installed version of pdelab fail miserably with unresolved symbols
when the library is checked for.

This patch solves this by adding the -ldunepdelab to the linker flags.
It should be applied to the release, of course. Otherwise testing _will_
fail.