[Bugfix][PkFEM] Make sure index compression for 2D FE variant selection is consistent
The PkFEM in 2D uses an array of 6 different FE variants for the
different possible permutations of the vertex indices. When looking up a
finite element, the subindices of the vertices are retrieved and
compressed to the range [0,5] depending on their permutation state.

Unfortunately, the code doing uses the inverse of the correct statement
for the relative ordering for vertices 1 and 2.

Fixed - thanks to Pavel for noticing!