[Communication] Change DataHandleProvider and EntityIndexCache to keep per-leaf size information
DataHandleProvider and EntityIndexCache cooperate using an informal
protocol to calculate the per-entity DOF count. The DataHandleProvider
obtains this information by visiting the leaves of the Ordering tree and
accumulating the size information at those leaf Orderings.

While the accumulated total is sufficient in the standard PDELab case,
MultiDomain requires the per-leaf information for certain combinations
of subdomain and processor boundaries, in particular if a subdomain
exists on the cells on one side of the processor boundary but not on the
other and the user has chosen a discretization with DOFs on entities
with codim > 0. In this case, the per-entity size information on the
different ranks can become inconsistent.

This patch changes the accumulation process to store the per-leaf sizes
in an array that is exported via the new method offsets() of the
EntityIndexCache.