[Communication] Add support for communicating entities with partially inconsistent DOF counts
This patch is required for supporting parallel computations with
MultiDomain, where it can be possible for entities of codim > 0 to have
different numbers of DOFs attached to them on different MPI ranks.

The problem arises if an entity belongs to a subdomain, but there is no
associated cell that also belongs to the subdomain on one MPI rank. In
that case, the GridFunctionSpace will not now about this membership and
thus consider the subspace assoicated with the subdomain to be empty,
causing a crash when too much / too few data is sent / received.

In order to fix this problem, there is now a new method sendLeafSizes()
on the DataHandleProvider. If that method returns true, the generic data
handle infrastructure will automatically communicate per-LeafOrdering
sizes and only enforce sizes to be consistent for those leaf
orderings. By default, this method always returns false, so most
existing applications should not even notice the difference between the
old and the new implementation.

Please note that this change requires wrapping the message buffer passed
in by the grid, as we now have to send heterogeneous data (e.g. sizes
and doubles) over the same data handle. This is not supported by the
standard data handles, so we use a DataHandle for char and marshal
everything to a simple byte stream. However, these changes only apply to
per-DOF communication. Per-entity communication does not require the
additional overhead, as there is no ambiguity as to the subdomain
membership of entities at the grid level.