[VariableMonomFEM] Check whether bound entities have correct GeometryType
VariableMonomFEM only supports a single GeometryType for grid cells,
even though the fact that you can pass in an arbitrary mapper seems to
suggest otherwise; the returned finite elements are only correct for
the GeometryType passed to the FEM's constructor or the default
GeometryType, which is a cube.

Unfortunately, the FEM did not check whether the grid cell passed to
find() actually had the correct GeometryType, leading to weird, subtle
bugs in the ordering infrastructure later on. This was noticed by
Christoph Grüninger, who ran testordering with assertions turned on,
promptly tripping over a check to make sure that the current
GeometryType is actually in use with the Ordering. Thanks for reporting,
Christoph!

This was really nasty to debug - it's fixed now by checking the
GeometryType in find(), but it's still possible to create a wrong FEM
for a given grid. After the release, we should change the FEM to simply
take the GridView in its constructor and automate the whole process,
maybe even adding support for different GeometryTypes. But for now, the
quick fix will have to do.