[Testordering] Initialize VariableMonomFEM with the correct GeometryType
VariableMonomFEM for a simplex ALUGrid was initialized with the default
constructor, which assumes a hexahedral grid, creating subtle
inconsistencies in the ordering data structures. These only caused a
test failure when compiled with assertions, which is why the problem
went unnoticed for that long.

Fixed by explicitly calling the correct constructor. Thanks to Christoph
Grüninger for reporting the original issue.