[Bugfix][GridFunctionSubSpace] Fix LFS specialization without template aliases
GridFunctionSubSpace lives in different namespaces depending on whether
the compiler supports template aliases or not. If it does, the class
lives in Dune::PDELab::gfs, otherwise there is a subclass in
Dune::PDELab. This difference screws up the LocalFunctionSpace
specialization in the backwards compatibility cases.

Fixed with a bunch of ugly preprocessor hacks.