Remove testpoisson-globalfe
The test is (as the codebase it is testing) suffering from bitrot.
To make "make check" pass, this test is either to be deleted or
be ignored. As it consists mostly of a duplication of testpoisson,
I propose to modify testpoisson to cover globalfe, as soon as the
globalfe code is working again.