[gridoperator][enginefactory] Introduce the abstraction of an engine factory.
An engine factory encapsulates how to obtain an engine from a hierarchy of
local assemblers.  This is needed for multithreaded assemblers:

Currently, the GridOperator obtaines a reference to the engine from the local
assembler and configures it, but it does not know when a new engine is needed
for a new thread.  On the other hand, the Assembler does know when a new
engine is needed, but it knows neither how to obtain it from the
LocalAssembler, nor does it know how to configure it.  The EngineFactory's job
is to encapsulate the information how to obtain and configure the engine from
a new local assembler.  The Engine factory is created by the GridOperator,
which has the necessary information, and handed to the global assembler, which
knows when to use it.