[unfriend][gridoperator] Unfriend the engines from the local assemblers.
Making the engines friends of the local assemblers was necessary so they could
query the state of the doPre/Postprocessing flags.  This had the disadvantage
that we could not easily derive one engine from another without the need to
make the base engine a fried as well.

This patch provides public interface methods to access the necessary
information.