[VTK] Automatically deduce function type for factory-based adding of custom gridfunctions
The OutputCollector returned from addSolutionToVTKWriter() supports
adding additional cell and vertex functions to the writer using a
factory-based approach. This patch changes those methods to use decltype
for deducing the exact type of the generated function and removes the
need for a custom typedef-based protocol between the OutputCollector and
the factory.