[ISTL] Fix Dirichlet constraints application for nested BCRS matrices
When using nested BCRS matrices in multidomain to implemented coupling
schemes like Neumann-Dirichlet, applying Dirichlet constraints to the
matrix fails if the associated diagonal block is missing (because it
belongs to the subproblem we are not interested in right now). In that
case, attempting to write a 1 on the diagonal raises an exception.

This patch fixes the problem by adding a new helper method
istl::write_matrix_element_if_exists(), which makes sure that the block
exists before traversing into it. This new method is used by the matrix
wrapper when setting Dirichlet constraints.