[BlockMatrixDiagonal][Bugfix] Fix construction of last-level row block end iterators
BlockMatrixDiagonal is a helper class for the parallel nonoverlapping
Jacobi implementation that provides optimized storage for the diagonal
matrix blocks, as those need to be communicated to create a consistent
preconditioner.

In order to create a generic infrastructure for setting up this data
structure, the setup procedure doesn't directly use the native iterators
of the block rows to copy the per-DOF data, but replaces them with
simple pointers to the dense, row-major data.

Unfortunately, the construction of the end iterators in this scheme
simply dereferenced the original end() iterator and then took the
address of that (non-existing) element. This is not legal C++, however,
and rightfully caused breakage when compiling with assertions.

This patch works around the problem by instead decreasing the end()
iterator by one, dereferencing that (valid) iterator, and incrementing
the address of the resulting element.

Kudos to Nagaiah Chamakuri for reporting the problem and to Dominic
Kempf for debugging it!

This patch fixes a problem related to FS#1549.