[Ordering] Add new ChunkedBlockOrdering decorator
This patch adds a new decorator ordering::Chunked that makes it possible
to introduce a "stupid" blocking scheme at any given function space
node. It must be given a chunk size and simply requires the block count
of the underlying ordering node to be a multiple of that chunk size.