[autotools] Fix build failure on conan in the absense of the suggested dune-alugrid.
One Makefile.am was using the automake conditional DUNE_ALUGRID, probably a
leftover from the old alugrid, or a borked attempt to migrate to the new
alugrid.  However the correct automake conditional is actually
HAVE_DUNE_ALUGRID for the new dune-alugrid module.
[autotools] Fix build failure on conan in the absense of the suggested dune-alugrid.

One Makefile.am was using the automake conditional DUNE_ALUGRID, probably a
leftover from the old alugrid, or a borked attempt to migrate to the new
alugrid.  However the correct automake conditional is actually
HAVE_DUNE_ALUGRID for the new dune-alugrid module.