[ICC][Workaround][PowerGFS] Work around compiler bug in PowerGFS constructor on ICC
ICC has a bug, where it fails to correctly construct a parameter pack
from a composite expresssion if that pack is the only template
parameter. This breaks one of the PowerGridFunctionSpace constructors.

This patch fixes the issue by breaking out the first argument of the
parameter pack into a separate template argument.