[ISTL] Move VectorBackend and parameter enum into istl namespace
Having the vector backend descriptor and the matrix backend descriptor
in different namespaces is just confusing, so this patch does
ISTLVectorBackend -> istl::VectorBackend. At the same time, the
ISTLParameters namespace was only needed due to the lack of scoped enums
in C++03. As we can now rely on those being available, this patch also
introduces a new enum istl::Blocking for the blocking types. It also
improves the naming to more clearly state what will happen by choosing
an option. The old enum is still there and has to be used with the old
backend. Both the enum and the old backend class have been deprecated.