[ISTL] Switch AMGDG backends to new istl::BCRSMatrixBackend
The AMGDG backends still use the old ISTLMatrixBackend with the slow
pattern creation. This patch switches them to the new backend. However,
that backend needs an estimate for the number of entries per matrix
row. The user can either use the old constructor, which will use a
heuristic to guess a value, or use a new constructor that takes a
ParameterTree instead of a list of parameters. In that ParameterTree,
the value for the number of entries per row can be specified with the
key "low_order_space.entries_per_row".