[Common] Add PartitionViewEntitySet
In Dune 2.4, GridViews for partitions other than AllPartition have been
deprecated. This is mainly due to the fact that their IndexSet does not
match the given partition, as GridViews always provide an IndexSet for
AllPartition.

On the other hand, PDELab currently uses a GridView on the
InteriorBorder partition for non-overlapping parallel computations,
where we have also suffered from the Indexset mismatch.

The deprecation has now given me a reason to properly solve the problem
by writing a new class that represents a view of a given set of
partitions of a GridView, together with an IndexSet for just those
partitions.

This patch adds the PartitionViewEntitySet to do just that. Its
interface is very similar to the GridView that it wraps, but there are
some differences:

- You have to explicitly tell the entity set about all codims that you
  intend to use with it. This is mainly an optimization, because you
  don't want to generate and store indices for all codimensions if the
  user is doing DG and only needs index information for the cells.

- As you cannot obtain an entity set from the grid, you have to manually
  construct it, which can be expensive. It is, however, cheap to copy
  entity sets around (internally, they only hold a shared_ptr to their
  IndexSet, which does all the heavy lifting.

- Their IndexSet obviously only spans the set of partitions that the
  entity set was constructed for and provides a proper index range for
  those entities.

- It will not report any ghost or overlap if those partitions are not
  part of the entity set, even though the underlying GridView might
  contain a ghost / overlap region.

- The IndexSet has a slightly extended interface: In addition to the
  member function types(codim), it also provides types(), which gives
  you the GeometryTypes of all codims at once, and new methods
  uniqueIndex() and uniqueSubIndex(), which return an index that is
  unique across all contained GeometryTypes. This exists mostly to avoid
  having to create an extra Mapper just to get at that information,
  which is already contained in the IndexSet anyway.

The PartitionViewEntitySet for the AllPartition is specialized and
directly forwards all index and size calls to the underlying GridView,
which avoids building the possibly expensive index map into the smaller
set of partitions.

Users usually should not directly create a PartitionViewIndexSet, but
use one of the provided template aliases AllEntitySet,
OverlappingEntitySet and NonOverlappingEntitySet, which take care of
picking the correct set of grid partitions.