[FiniteElementMap] Add new interface method hasDOFs(codim)
For the new EntitySet interface, we have to be able to determine which
codimensions to add to the EntitySet. For this purpose, all
FiniteElementMaps now provide a method hasDOFs(codim), which must return
true if there is any chance that the given FiniteElementMap will attach
data to an entity of that codimension.

Important: Note that this also holds for *variable-sized* FEMs. While
those FEMs are allowed to throw an exception on any call to
size(GeometryType), they have to return a correct value from hasDOFs().