[GridFunctionSpace] Use EntitySet instead of GridView
This patch ports the function spaces to internally always use an
EntitySet instead of a GridView. The leaf space (the only place where
users can specify a GridView), now detects whether the user passed in an
EntitySet or a GridView; in the latter case, the GridView is
automatically wrapped in an AllEntitySet which is then used internally.