[PartitionViewEntitySet] Make index set construction work for grids which lack entities
Not all grids offer entities (or, more importantly, iterators) for all
codimensions, making it impossible to use our simple approach of
iterating over all entities of all codimensions (UG, I'm looking at
you...). On the other hand, we still have to provide index information
for all codimensions. So this patch adds a second construction method
that only iterates over the cells and obtains all other indices using
the subIndex() method. This is probably a lot slower, but it works...