[GridFunctionSpace] Port VectorGridFunctionSpace to EntitySet
When porting the function spaces, I forgot the
VectorGridFunctionSpace. This patch plugs that hole.

As VectorGridFunctionSpace is also passed the GridView, it needs to do
the backwards compatibility dance of the regular leaf space. In order to
avoid writing those conditionals in a bunch of places, I moved them to
template aliases in the impl namespace, defined in gridfunctionspace.hh.