[Tests] Port testnonoverlapping to the new API
The test now runs on both a simplex and a cube grid, and the cube grid
is constructed using the GridFactory to avoid external grid files - DGF
begone!

Moreover, the test is now configured to run on 2 MPI ranks in parallel,
which tests the concave corner problem for which the nonoverlapping
GridOperator has special handling.