[Assembler] Upgrade IntersectionType helper for EntitySets
The IntersectionType helper from the GridOperator utilities is used
inside the GridOperator assembler to classify the type of an
intersection (boundary, skeleton etc.).

This patch updates the IntersectionType classifier to correctly handle
the fact that an EntitySet might not span the whole underlying
GridView. As this extended test will often require access to the outside
entity, the new classifier function returns the outside entity along
with the IntersectionType.

The patch also moves the whole classifier machinery to a separate
header, as the constraints assembler (which is pretty broken right now)
will also have to use it in the future.