[Cleanup] Deprecate PermutationOrdering
PermutationOrdering is a combination of a lexicographic ordering with a
permutation step that is implemented for Power and Composite spaces. It
has been superceded by the more flexibe Permuted<> decorator that can be
applied to arbitrary orderings.

(cherry picked from commit f801c549ed7b01cc8918a59fe4ff0417bdee5041)
Signed-off-by: Steffen Müthing <muething@dune-project.org>