[EntitySet] Allow users to force an update
EntitySet aggressively caches the "freshness" of its internal state and
short-circuits update() calls if it determines that its state hasn't
changed since the last call to update().

Unfortunately, this logic fails if the underlying GridView has changed,
e.g. due to grid adaptivity. This patch adds a bool parameter to
update() that allows the user to force an update to happen.