use lop gather and scatter method in gridoperator
Note: This only tests the infrastsructure. The residual assembly should
be right but the jacobian isn't doing anything useful. That is no
problem since we want to use a matrix free solver in the end and this
needs only jacobian_apply methods which correspond to alpha methods from
the local operator.