[CMake][bugfix] Only run the complex tests with UG present
The code base as it is throws a `#error` if UG is not present.
That is _NOT_ a nice behaviour for a test. This commit applies
the simple fix of only adding the test with UG present. It would
of course be nicer to also check complex number computations with
YaspGrid, but that is much more work.