[Backend] Add aliased vector and matrix views
These new views directly alias into a corresponding block of the global
vector / matrix. They require the associated storage in the global
vector / matrix to be contiguous and can be used as a drop-in
replacement for LocalVector and LocalMatrix.

As an additional API, they have a member data() that exposes a pointer
to the underlying storage, which is e.g. used in the sumfactorized local
operators as input to the sumfactorization kernels.