[LocalFunctionSpace] Add block-based bind() variant
By default, the LFS will now use a block-based approach to bind():

- Only a single DOFIndex (the first one) gets mapped
- The FiniteElement is not updated during bind()

This only works correctly for fixed-size DG-based spaces that use
cell-wise blocking in the linear algebra! the traditional way of binding
can be requested by passing an instance of std::false_type as the second
parameter to bind().