Modify jacobian_boundary to allow usage of CCFV operator for rediscretisation in AMG4DG
If the operator is used to discretise the coarse grid problem in AMG4DG directly, calculate the distance to the centre of the boundary grid cell, since this is done in the DG treatment of boundary conditions. If the operator is also multiplied by alpha*p*(p-1+dim), the operator agrees with what is obtained via the triple matrix-product.