[CG2DG,Backport] take implementation from exadune-appl
I never liked the old implementation based on the `ComputeCG2DGVisitor`
because it seems not straightforward to me.
The implementation from exadune-appl directly calculates the
prolongation from a matrix-matrix product. It needs the same
amount of work as the old implementation in dune-pdelab and has the
following pros:

- alpha_volume-method straightforward from the formula, has been
  implemented
- alpha_volume of restriction also straightforward, useful for
  matrix-free computations, will be implemented when it is needed

The adopted implementation from exadune-appl has been modified in order
to minimize the amount of data copying.

This change can be backported to master.