[Utility][Backport] Add helper functions for obtaining reference elements and quadrature rules
This patch adds freestanding functions quadratureRule() and
referenceElement() to PDELab that obtain the corresponding objects given
a geometry. This avoids lots of boilerplate code (extraction of ctype
and dimension etc.). In addition, this patch also provides wrappers for
reference elements and quadrature rules that have value semantics,
allowing users to capture the result of the new functions using plain
auto.

Finally, the new functions are also injected into the Dune namespace to
enable unqualified ADL lookup. This is more convenient and more
extensible and also provides easy forward compatibility once the
functions make their (planned) appearance in the core modules.