[CMake] Check for C++14
This commit adds a check for C++14 and switches PDELab and all dependent
modules to C++14 mode, if available. By default, this is entirely
optional, and PDELab will work fine in C++11 mode. There will be some
optional components that are only available in C++14, though.

If you module absolutely requires these optional components, you can
call the function pdelab_require_cxx14() in the main CMakeLists.txt of
your module. That function will raise a fatal error during configuration
if C++14 is not supported.

Note: Due to the minimal level of C++14 support in older compilers, this
version only checks for -std=c++14 and not for -std=c++1y. This is in
contrast to the check in the core modules of Dune 3.0.