[GridOperator] Extend callswitch to handle nonlinear jacobian_apply() methods
The new nonlinear variant of jacobian_apply() introduces an additional
set of jacobian_apply() methods, which requires a new set of callswitch
guards as added by this patch.

Moreover, the new methods differ from the existing methods by an
additional parameter that specifies the linearization point for the
Jacobian. As the user might not necessarily have implemented both
variants in the future, we should make sure to provide understandable
error messages. So this patch also adds a few static_asserts() that tell
the user about the missing methods. These asserts are built on top of
the new concept framework in dune-common. As such, backporting this will
require pulling the concept framework into PDELab for 2.4.